fix: naver OAuth2 join process
