+++ src/main/java/org/example/honorsparkingbe/domain/entity/MemberEntity.java
@@ -21,32 +21,32 @@ public class MemberEntity {
     private CarEntity carEntity;
 
     @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
+    // @Column(nullable = false)
     private LoginPlatform loginPlatform;
 
     @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
+    // @Column(nullable = false)
     private MemberRole role;
 
-    @Column(nullable = false, length = 60)
+    // @Column(nullable = false, length = 60)
     private String userName;
 
-    @Column(nullable = false, unique = true)
+    // @Column(nullable = false, unique = true)
     private String authId;
 
-    @Column(nullable = false)
+    // @Column(nullable = false)
     private String password;
 
-    @Column(nullable = false, length = 15)
+    // @Column(nullable = false, length = 15)
     private String phoneNumber;
 
-    @Column(nullable = false, unique = true)
+    // @Column(nullable = false, unique = true)
     private String email;
 
-    @Column(nullable = false)
+    // @Column(nullable = false)
     private int birthdayYear; // 생년 (정수형)
 
-    @Column(nullable = false, length = 4)
+    // @Column(nullable = false, length = 4)
     private String birthday; // 생일 (문자열, "MMDD" 형태로 저장할 수 있음)
 
 }
